USGS Lidar Toolkit -- Prepare Target Data

Prepare features for use when querying a lidar project or tile index to determine available data coverage. The basic operation done by this function is to apply the buffer to the features.

prepareTargetData(
  x,
  y,
  buffer = 0,
  shape = "square",
  aoi = "",
  crs = "",
  segments = 60,
  returnType = "Spatial"
)

Arguments

x
Location or list of locations containing the easting for the center of the area-of-interest.

y
Location or list of locations containing the northing for the center of the area-of-interest.

buffer
Distance or list of distances added or subtracted to (x,y) to create the area-of-interest. Can be 0 when used with (x,y) to return projects providing coverage for point location(s). Can be a vector of values corresponding to the number of objects in aoi allowing a different buffer size for each object when aoi is a Spatial* of sf object. buffer can be negative with polygon feature to reduce the area considered. However, you may end up with weird shapes if the distance is larger than the width of the polygon.

shape
Character string describing the shape of the sample area. Valid values are "square" or "circle". The buffer shape is the same for all objects specified by (x,y) or aoi.

aoi
Spatial* or sf object containing a point(s) or polygon(s) describing the area(s) of interest. Can be points or polygons.

crs
Valid proj4string string defining the coordinate reference system of (x,y) or aoi. crs is required when using (x,y). It is optional when using aoi. This can also be a valid SRS_string for use with the sp::CRS() function where it will be provided as the projargs argument.

segments
Number of segments to use when generating a circular area of interest. When using a SpatialPoint* or sf object with shape = "circle", set segments to a rather large value (60 or higher) that is a multiple of 4. The gBuffer function from rgeos and st_buffer function from sf are used to build the sample areas and it accepts the number of segments in a quarter circle so small values for segments may not produce good circles. Values for segments that are not a multiple of 4 will not produce circles with the correct number of segments.

returnType
Character string specifying the object type for the polygon object returned by prepareTargetData when (x,y) is used to specify the area-of-interest. Valid values are "Spatial" or "sf". returnType is ignored when aoi is specified and the return type with match the type of aoi.

Value

A set of optionally buffered spatial features. The return type will be the same as the aoi type. When (x,y) is used, returnType specifies the object type returned by prepareTargetData.

Details

Prepare point or polygon data for use when querying to determine the lidar project or tile coverage. Typically this preparation is done by the queryUSGSProjectIndex or queryUSGSTileIndex functions but you can use prepareTargetData to build the query objects so they can be used for other purposes.

Units for buffer are the same as the horizontal units for the input features so you shouldn't use data in LON-LAT with units of degrees.

Examples

prepareTargetData(-13540901, 5806426, 180, shape = "circle",
  crs = sp::CRS(SRS_string="EPSG:3857")@projargs)

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs

#> Warning: Discarded datum World Geodetic System 1984 in Proj4 definition

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs

#> Warning: Discarded datum World Geodetic System 1984 in Proj4 definition

#> An object of class "SpatialPolygonsDataFrame"
#> Slot "data":
#>   ID
#> 1 P1
#> 
#> Slot "polygons":
#> [[1]]
#> An object of class "Polygons"
#> Slot "Polygons":
#> [[1]]
#> An object of class "Polygon"
#> Slot "labpt":
#> [1] -13540901   5806426
#> 
#> Slot "area":
#> [1] 101601.7
#> 
#> Slot "hole":
#> [1] FALSE
#> 
#> Slot "ringDir":
#> [1] 1
#> 
#> Slot "coords":
#>               x       y
#>  [1,] -13540721 5806426
#>  [2,] -13540722 5806407
#>  [3,] -13540725 5806389
#>  [4,] -13540730 5806370
#>  [5,] -13540737 5806353
#>  [6,] -13540745 5806336
#>  [7,] -13540755 5806320
#>  [8,] -13540767 5806306
#>  [9,] -13540781 5806292
#> [10,] -13540795 5806280
#> [11,] -13540811 5806270
#> [12,] -13540828 5806262
#> [13,] -13540845 5806255
#> [14,] -13540864 5806250
#> [15,] -13540882 5806247
#> [16,] -13540901 5806246
#> [17,] -13540920 5806247
#> [18,] -13540938 5806250
#> [19,] -13540957 5806255
#> [20,] -13540974 5806262
#> [21,] -13540991 5806270
#> [22,] -13541007 5806280
#> [23,] -13541021 5806292
#> [24,] -13541035 5806306
#> [25,] -13541047 5806320
#> [26,] -13541057 5806336
#> [27,] -13541065 5806353
#> [28,] -13541072 5806370
#> [29,] -13541077 5806389
#> [30,] -13541080 5806407
#> [31,] -13541081 5806426
#> [32,] -13541080 5806445
#> [33,] -13541077 5806463
#> [34,] -13541072 5806482
#> [35,] -13541065 5806499
#> [36,] -13541057 5806516
#> [37,] -13541047 5806532
#> [38,] -13541035 5806546
#> [39,] -13541021 5806560
#> [40,] -13541007 5806572
#> [41,] -13540991 5806582
#> [42,] -13540974 5806590
#> [43,] -13540957 5806597
#> [44,] -13540938 5806602
#> [45,] -13540920 5806605
#> [46,] -13540901 5806606
#> [47,] -13540882 5806605
#> [48,] -13540864 5806602
#> [49,] -13540845 5806597
#> [50,] -13540828 5806590
#> [51,] -13540811 5806582
#> [52,] -13540795 5806572
#> [53,] -13540781 5806560
#> [54,] -13540767 5806546
#> [55,] -13540755 5806532
#> [56,] -13540745 5806516
#> [57,] -13540737 5806499
#> [58,] -13540730 5806482
#> [59,] -13540725 5806463
#> [60,] -13540722 5806445
#> [61,] -13540721 5806426
#> 
#> 
#> 
#> Slot "plotOrder":
#> [1] 1
#> 
#> Slot "labpt":
#> [1] -13540901   5806426
#> 
#> Slot "ID":
#> [1] "1"
#> 
#> Slot "area":
#> [1] 101601.7
#> 
#> 
#> 
#> Slot "plotOrder":
#> [1] 1
#> 
#> Slot "bbox":
#>         min       max
#> x -13541081 -13540721
#> y   5806246   5806606
#> 
#> Slot "proj4string":
#> CRS arguments:
#>  +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1
#> +units=m +nadgrids=@null +wktext +no_defs 
#> 

pt1 <- sf::st_point(c(-13540901, 5806426 + 500))
pt2 <- sf::st_point(c(-13540901 + 500, 5806426 - 500))
pt3 <- sf::st_point(c(-13540901 - 500, 5806426))
pt4 <- sf::st_point(c(-13540901 + 1000, 5806426 - 1000))
id <- c("P1", "P2", "P3", "P4")
x_sf <- sf::st_sf(data.frame(ID = id, stringsAsFactors = FALSE),
  geom = sf::st_sfc(pt1, pt2, pt3, pt4),
  crs = sp::CRS(SRS_string="EPSG:3857")@projargs)

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs

#> Warning: Discarded datum World Geodetic System 1984 in Proj4 definition

samples <- prepareTargetData(aoi = x_sf, buffer = 180)