USGS Lidar Toolkit -- Identify USGS Lidar Tiles Covering Area

Intersect a set of features (points or polygons) against a lidar tile index to determine which tiles are needed to provide coverage for the features.

queryUSGSTileIndex(
  x,
  y,
  buffer = 0,
  projectID = "",
  fieldname = "project_id",
  shape = "square",
  aoi = NA,
  crs = "",
  index = "",
  segments = 60,
  return = "index",
  returnType = "Spatial",
  returncrs = "same",
  ...
)

Arguments

x
Location easting for the center of the area-of-interest.

y
Location northing for the center of the area-of-interest.

buffer
Distance in meters added or subtracted to (x,y) to create the area-of-interest. Can be 0 when used with (x,y) to return items providing coverage for point location(s). Can be a vector of values corresponding to the number of objects in aoi allowing a different buffer size for each object when aoi is a Spatial* or sf object. buffer can be negative with polygon features to reduce the area considered. However, you may end up with weird shapes if the distance is larger than the width of the polygon. Data in (x,y) and aoi are projected into the web mercator coordinate reference system before the buffer operation is done so the units for buffer are always meters.

projectID
Character string or list of character strings containing the ID(s) for the lidar project(s). Typically obtained by calling queryUSGSProjectIndex().

fieldname
Character string containing the name of the field to use when querying the tile index.

shape
Character string describing the shape of the sample area in the case of point features or the shape applied to the buffer corners when using polygon features. Valid values are "square" or "circle".

aoi
Spatial* or sf object containing a point or polygon describing the area of interest. Can be points or polygons.

crs
Valid proj4string string defining the coordinate reference system of (x,y). crs is required when using (x,y). crs is ignored when aoi is specified.

index
Index file for USGS lidar tiles If not provided, an index previously specified by a call to fetchUSGSProjectIndex or setUSGSProjectIndex will be used. If not provided and you have not called fetchUSGSTileIndex or setUSGSTileIndex, the function displays a warning and returns.

segments
Number of segments to use when generating a circular area of interest. When using a SpatialPoint* or sf object with shape = "circle", set segments to a rather large value (60 or higher) that is a multiple of 4. The gBuffer function from rgeos and st_buffer function from sf are used to build the sample areas and it accepts the number of segments in a quarter circle so small values for segments may not produce good circles. Values for segments that are not a multiple of 4 will not produce circles with the correct number of segments.

return
Character string specifying the kind of shapes to be returned. Valid values are ""aoi" to return shapes for each object in the aoi or the object defined by (x,y) and buffer with additional attributes providing lidar project information and "index" (default) to return tile shapes that provide coverage for the features in the "aoi" or the area defined by (x,y) and buffer. When tile shapes are returned there is no "aoi" information included with the project shape attributes.

returnType
Character string specifying the object type for the returned polygon object when (x,y) is used to specify the area-of-interest. Valid values are "Spatial" or "sf". returnType is ignored when aoi is specified.

returncrs
Valid proj4string string defining the coordinate reference system of the returned SpatialPolygonsDataFrame or sf object. A value of "same" will project the return object to the same coordinate reference system as aoi or to crs when used with (x,y).

...
Additional arguments passed to download.file

Value

A SpatialPolygonsDataFrame or sf object containing polygon(s) and attribute(s) for tiles covering the specified area. Attributes for the aoi are placed first in the data.frame followed by attributes for the lidar tile polygons..

Details

Query the tile index to find tiles that intersect the buffered (x,y) point(s) or buffered feature(s) provided in aoi.

Examples

if (FALSE) {
queryUSGSTileIndex(-13540901, 5806426, 180, shape = "circle",
  CRS = CRS(SRS_string="EPSG:3857"))
  }